UI changes - progress bar - disconnected
Flipped progress bar on X axis so it fills rather than empties.

Changed color of progress bar - however since it is flipped on the X the background color becomes the fill color, which is a bit faint.

I added all of the code for the disconnected symbol as well as making the main text invisible while disconnected is displayed, but in the old method it was called from init and update, and we have neither in that fragment. I am unsure of where to put them.

Further, the main text of the screen becomes invisible when Disconnected activates, except use room now is called by TrafficLightsPresenter which overrides what I am trying to do. I think I have a solution because perhaps someone else would like to implement something cleaner than my hack and slash approach